`xtask`: Outsource handling of compilation targets into `mod targets`
